scraper/requirements.txt
requests
beautifulsoup4
pytz
python-dateutil
